modules/training_pipeline/requirements.txt
accelerate==0.21.0 ; python_version >= "3.10" and python_version < "3.11"
aiohttp==3.8.5 ; python_version >= "3.10" and python_version < "3.11"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "3.11"
astor==0.8.1 ; python_version >= "3.10" and python_version < "3.11"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "3.11"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "3.11"
backoff==2.2.1 ; python_version >= "3.10" and python_version < "3.11"
beam-sdk==0.15.9 ; python_version >= "3.10" and python_version < "3.11"
beautifulsoup4==4.12.2 ; python_version >= "3.10" and python_version < "3.11"
bitsandbytes==0.41.1 ; python_version >= "3.10" and python_version < "3.11"
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "3.11"
chardet==5.2.0 ; python_version >= "3.10" and python_version < "3.11"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "3.11"
click==8.1.6 ; python_version >= "3.10" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.11" and platform_system == "Windows"
comet-llm==1.1.0 ; python_version >= "3.10" and python_version < "3.11"
comet-ml==3.35.1 ; python_version >= "3.10" and python_version < "3.11"
configobj==5.0.8 ; python_version >= "3.10" and python_version < "3.11"
croniter==1.4.1 ; python_version >= "3.10" and python_version < "3.11"
dataclasses-json==0.6.1 ; python_version >= "3.10" and python_version < "3.11"
datasets==2.14.3 ; python_version >= "3.10" and python_version < "3.11"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "3.11"
dill==0.3.7 ; python_version >= "3.10" and python_version < "3.11"
dulwich==0.21.5 ; python_version >= "3.10" and python_version < "3.11"
einops==0.6.1 ; python_version >= "3.10" and python_version < "3.11"
emoji==2.8.0 ; python_version >= "3.10" and python_version < "3.11"
everett[ini]==3.1.0 ; python_version >= "3.10" and python_version < "3.11"
filelock==3.12.2 ; python_version >= "3.10" and python_version < "3.11"
filetype==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
fire==0.5.0 ; python_version >= "3.10" and python_version < "3.11"
flatten-dict==0.4.2 ; python_version >= "3.10" and python_version < "3.11"
frozenlist==1.4.0 ; python_version >= "3.10" and python_version < "3.11"
fsspec==2023.6.0 ; python_version >= "3.10" and python_version < "3.11"
fsspec[http]==2023.6.0 ; python_version >= "3.10" and python_version < "3.11"
huggingface-hub==0.16.4 ; python_version >= "3.10" and python_version < "3.11"
idna==3.4 ; python_version >= "3.10" and python_version < "3.11"
importlib-metadata==5.2.0 ; python_version >= "3.10" and python_version < "3.11"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "3.11"
joblib==1.3.2 ; python_version >= "3.10" and python_version < "3.11"
jsonschema-specifications==2023.7.1 ; python_version >= "3.10" and python_version < "3.11"
jsonschema==4.19.0 ; python_version >= "3.10" and python_version < "3.11"
langdetect==1.0.9 ; python_version >= "3.10" and python_version < "3.11"
lxml==4.9.3 ; python_version >= "3.10" and python_version < "3.11"
markdown-it-py==3.0.0 ; python_version >= "3.10" and python_version < "3.11"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "3.11"
marshmallow-dataclass==8.5.14 ; python_version >= "3.10" and python_version < "3.11"
marshmallow==3.18.0 ; python_version >= "3.10" and python_version < "3.11"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "3.11"
mpmath==1.3.0 ; python_version >= "3.10" and python_version < "3.11"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "3.11"
multiprocess==0.70.15 ; python_version >= "3.10" and python_version < "3.11"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
networkx==3.1 ; python_version >= "3.10" and python_version < "3.11"
nltk==3.8.1 ; python_version >= "3.10" and python_version < "3.11"
numpy==1.25.2 ; python_version >= "3.10" and python_version < "3.11"
packaging==23.1 ; python_version >= "3.10" and python_version < "3.11"
pandas==2.0.3 ; python_version >= "3.10" and python_version < "3.11"
peft==0.4.0 ; python_version >= "3.10" and python_version < "3.11"
psutil==5.9.5 ; python_version >= "3.10" and python_version < "3.11"
pyarrow==12.0.1 ; python_version >= "3.10" and python_version < "3.11"
pygments==2.16.1 ; python_version >= "3.10" and python_version < "3.11"
python-box==6.1.0 ; python_version >= "3.10" and python_version < "3.11"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "3.11"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
python-iso639==2023.6.15 ; python_version >= "3.10" and python_version < "3.11"
python-magic==0.4.27 ; python_version >= "3.10" and python_version < "3.11"
pytz==2023.3 ; python_version >= "3.10" and python_version < "3.11"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.11"
rapidfuzz==3.4.0 ; python_version >= "3.10" and python_version < "3.11"
referencing==0.30.2 ; python_version >= "3.10" and python_version < "3.11"
regex==2023.6.3 ; python_version >= "3.10" and python_version < "3.11"
requests-toolbelt==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
requests==2.31.0 ; python_version >= "3.10" and python_version < "3.11"
rich==13.5.2 ; python_version >= "3.10" and python_version < "3.11"
rpds-py==0.9.2 ; python_version >= "3.10" and python_version < "3.11"
safetensors==0.3.1 ; python_version >= "3.10" and python_version < "3.11"
scipy==1.11.1 ; python_version >= "3.10" and python_version < "3.11"
semantic-version==2.10.0 ; python_version >= "3.10" and python_version < "3.11"
sentry-sdk==1.29.2 ; python_version >= "3.10" and python_version < "3.11"
simplejson==3.19.1 ; python_version >= "3.10" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.11"
soupsieve==2.4.1 ; python_version >= "3.10" and python_version < "3.11"
sympy==1.12 ; python_version >= "3.10" and python_version < "3.11"
tabulate==0.9.0 ; python_version >= "3.10" and python_version < "3.11"
termcolor==2.3.0 ; python_version >= "3.10" and python_version < "3.11"
tokenizers==0.13.3 ; python_version >= "3.10" and python_version < "3.11"
torch==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
tqdm==4.65.0 ; python_version >= "3.10" and python_version < "3.11"
transformers==4.31.0 ; python_version >= "3.10" and python_version < "3.11"
trl==0.4.7 ; python_version >= "3.10" and python_version < "3.11"
typeguard==2.13.3 ; python_version >= "3.10" and python_version < "3.11"
types-requests==2.31.0.2 ; python_version >= "3.10" and python_version < "3.11"
types-urllib3==1.26.25.14 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "3.11"
typing-inspect==0.9.0 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "3.11"
unstructured==0.10.28 ; python_version >= "3.10" and python_version < "3.11"
urllib3==2.0.4 ; python_version >= "3.10" and python_version < "3.11"
validators==0.20.0 ; python_version >= "3.10" and python_version < "3.11"
websocket-client==1.3.3 ; python_version >= "3.10" and python_version < "3.11"
wrapt==1.15.0 ; python_version >= "3.10" and python_version < "3.11"
wurlitzer==3.0.3 ; python_version >= "3.10" and python_version < "3.11"
xxhash==3.3.0 ; python_version >= "3.10" and python_version < "3.11"
yarl==1.9.2 ; python_version >= "3.10" and python_version < "3.11"
zipp==3.16.2 ; python_version >= "3.10" and python_version < "3.11"
